alerta de eliminar

+++ CRUD_MYSQLI-clientes/Index.php
@@ -71,7 +71,7 @@
                     <th scope="row">
                         <!--php?ID se conecta con editardatophp en $_REQUEST['ID']; $fila['id'] se conecta con la columna id de la db sys_can-->
                         <a href="Forms/EditarDato.php?ID=<?php echo $fila['idcliente']?>" class="btn text-white text-center bg-secondary">Editar Datos</a>
-                        <a href="CRUD/EliminarDato.php?ID=<?php echo $fila['idcliente']?>" class="btn btn-danger">Eliminar Datos</a>
+                        <a href="CRUD/EliminarDato.php?ID=<?php echo $fila['idcliente']?>"  onclick="return  confirm ('ESTA SEGURO'); false "class="btn btn-danger">Eliminar Datos</a>
                     </th>
                 </tr>
                 <?php
